requirements.txt
celery==5.3.6
Django==5.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.1
